+++ internal/datatransfer/fileexport.go
@@ -1,4 +1,4 @@
-package export
+package datatransfer
 
 import (
 	"compress/gzip"