+++ internal/handlers/handlers.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"db-portal/internal/config"
-	"db-portal/internal/export"
+	"db-portal/internal/datatransfer"
 	"db-portal/internal/internaldb"
 	"time"
 )
@@ -11,7 +11,7 @@ type Services struct {
 	Store           *internaldb.Store
 	CommandsConfig  *config.Config[config.CommandsConfig]
 	ServerConfig    *config.Config[config.Server]
-	Exporter        export.Exporter
+	Exporter        datatransfer.Exporter
 	JWTSecretKey    string
 	clockResolution time.Duration
 }