+++ internal/handlers/query.go
@@ -2,15 +2,17 @@ package handlers
 
 import (
 	"context"
-	"db-portal/internal/auth"
 	"db-portal/internal/db"
 	"db-portal/internal/response"
+	"db-portal/internal/security"
 	"fmt"
 	"net/http"
+	//"strings"
 )
 
 func (s *Services) QueryHandler(w http.ResponseWriter, r *http.Request) {
-	username := r.Context().Value(auth.UserContextKey).(string)
+
+	username := r.Context().Value(security.UserContextKey).(string)
 	conname := r.FormValue("conn")
 
 	// reload config files if needed
@@ -19,7 +21,7 @@ func (s *Services) QueryHandler(w http.ResponseWriter, r *http.Request) {
 	// get connection details
 	connDetails, err := s.Store.FetchConn(username, conname)
 	if err != nil {
-		http.Error(w, fmt.Sprintf("connection %v not found or not allowed", conname), http.StatusNotFound)
+		http.Error(w, fmt.Sprintf("connection <%v> not found or not allowed for user <%v>", conname, username), http.StatusNotFound)
 		return
 	}